bcachefs: Replace BUG_ON() with real error checking
Not adding the repair code just yet

+++ fs/bcachefs/btree_gc.h
@@ -11,7 +11,7 @@ void bch2_gc_thread_stop(struct bch_fs *);
 int bch2_gc_thread_start(struct bch_fs *);
 int bch2_initial_gc(struct bch_fs *, struct list_head *);
 u8 bch2_btree_key_recalc_oldest_gen(struct bch_fs *, struct bkey_s_c);
-u8 bch2_btree_mark_key_initial(struct bch_fs *, enum bkey_type,
+int bch2_btree_mark_key_initial(struct bch_fs *, enum bkey_type,
 				struct bkey_s_c);
 void bch2_mark_dev_metadata(struct bch_fs *, struct bch_dev *);
